% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateme-package.R
\docType{package}
\name{updateme-package}
\alias{updateme}
\alias{updateme-package}
\title{updateme: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
