% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateme_on.R
\name{updateme_on}
\alias{updateme_on}
\alias{updateme_off}
\title{Temporarily turn {updateme} off for all packages}
\usage{
updateme_on()

updateme_off()
}
\value{
\code{NULL}, invisibly
}
\description{
These functions may be useful if you find the slight delay caused by
{updateme} looking up new package versions to be slowing down your workflow.
}
\examples{
if (FALSE) {
  updateme_off()
  updateme_on()
}
}
\seealso{
\code{\link[=updateme_set_sources]{updateme_set_sources()}} to turn {updateme} off for individual
packages
}
