% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateme_set_sources.R
\name{updateme_sources_set}
\alias{updateme_sources_set}
\title{Configure {updateme} lookup of new package versions}
\usage{
updateme_sources_set(...)
}
\arguments{
\item{...}{Named or unnamed arguments. Values should be either:
\itemize{
\item \code{"github"}: new versions will be found from the package development
version on GitHub, if a repo can be identified using the package
\code{DESCRIPTION}
\item One of \code{names(getOption("repo"))}: latest versions will be taken from
this source, if available
\item A URL pointing to a GitHub repo, e.g.
\code{"https://github.com/wurli/updateme"}: the latest version \emph{for this
particular package} will be taken from this project
}

If arguments are named, names should indicate package which the option
should apply to. If unnamed, the option will apply to all packages. See
examples for more information.}
}
\value{
The result of setting
\verb{options(updateme.sources = <new_options>)}
}
\description{
This function is a helper for setting the \code{"updateme.sources"}
global option. It provides a user-friendly interface and validation of the
options you set.
}
\examples{
if (FALSE) {

  # If you want to check non-standard repos for new versions of packages,
  # you'll first have to set the repos global option. Note that each
  # option must be named for compatibility with {updateme}
  options(repos = c(

    # Your default repos, e.g. c(CRAN = "https://cloud.r-project.org")
    getOption("repos"),

    # The tidyverse r-universe, including dev versions of tidyverse packages
    tidyverse = "https://tidyverse.r-universe.dev",

    # The r-lib r-universe, including dev versions of infrastructure packages
    # like {cli}, {rlang}, etc
    rlib = "https://r-lib.r-universe.dev"
  ))

  # 1. New versions will first be looked up in the tidyverse r-universe
  # 2. If not found, they will be looked up from your usual CRAN mirror
  # 3. {bslib} will always be first looked up from GitHub
  # 4. {cli} will always be first looked up from the r-lib r-universe
  updateme_sources_set(
    "tidyverse",
    "CRAN",
    bslib = "https://github.com/rstudio/bslib", # Name is optional here
    cli = "rlib"
  )
}

}
