% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateme-package.R
\docType{package}
\name{updateme-package}
\alias{updateme}
\alias{updateme-package}
\title{updateme: Informative Messages About Outdated Packages}
\description{
When a package is loaded, the source repository is checked for new versions and a message is shown in the console indicating whether the package is out of date.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wurli/updateme}
  \item \url{https://wurli.github.io/updateme/}
  \item Report bugs at \url{https://github.com/wurli/updateme/issues}
}

}
\author{
\strong{Maintainer}: Jacob Scott \email{jscott2718@gmail.com}

}
\keyword{internal}
