% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateme_on.R
\name{updateme_on}
\alias{updateme_on}
\alias{updateme_off}
\title{Temporarily turn updateme off for all packages}
\usage{
updateme_on()

updateme_off()
}
\value{
The old \code{updateme.on} option, as returned by \code{options()}
}
\description{
These functions may be useful if you find the slight delay caused by
updateme looking up new package versions to be slowing down your workflow.
}
\seealso{
\code{\link[=updateme_sources_set]{updateme_sources_set()}} to turn updateme off for individual
packages
}
